Fill out  buildWeatherLocation() in WeatherEntry.
This commit fills out the "build this Uri" method that was left unimplemented in Google's starter ContentProvider code. I'm not sure if having these methods in this place (the contract subclasses) is standard code practice, but I understand what they're doing. This commit also uncomments the testBuildWeatherLocation() method in TestWeatherContract, which is a test for the method I filled out.