Add code in FetchWeatherTask to insert OWM data into the DB.
This is the first part of using the database and ContentProvider fully. FetchWeatherTask now parses the JSON API response and uses the provider's bulkInsert() to put the data into the database. Still TODO is pulling out this data for use in the UI; right now the old "just parse the JSON directly into strings" code is still what is being used in the UI.