Use database projection in ForecastFragment.
Note: this does not do everything that the corresponding Udacity commit does. They additionally put a bunch of constant int "indices" in ForecastFragment to "speed up" the convertCursorRowToUXFormat(...) in ForecastAdapter, with each of these int constants referring to a row from the String[] projection. However, the way they do it doesn't look good to me. The String[] and the ints have no formal relationship, and using the ints in ForecastAdapter is kind of side-effect-y in that it assumes correctness of both the ints/String[] relationship, and the fact that it is indeed working with a cursor that was fed the String[] projection. This is bad because neither the String[]/int relationship, nor the ability to tell whether the cursor was fed the correct projection String[], is provable. At least, the way they did it in their commit for this lesson does not prove either of those, and just works on them as assumptions.