Made DetailView work again. Mostly.
The text that DetailView shows the user right now is a semi-placeholder. I believe this is a correct way to use a Loader and update a TextView in that activity/fragment, but I just want to move on to the next step in the course and they probably are going to do it some weird way that I'll have to change to anyway.